Move renderer `act` to work loop
